+++ package.json
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/desertgazelle/pluralsight-node-api#readme",
   "dependencies": {
+    "body-parser": "^1.20.0",
     "express": "^4.17.3",
     "mongoose": "^6.2.11",
     "nodemon": "^2.0.15"
